requirement.txt
certifi==2022.9.24
charset-normalizer==2.1.1
config==0.5.1
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
greenlet==2.0.1
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
mysql-connector-python==8.0.31
networkx==2.8.8
numpy==1.23.4
packaging==21.3
panda==0.3.1
pandas==1.5.1
Pillow==9.3.0
protobuf==3.20.1
psycopg==3.1.4
psycopg-binary==3.1.4
psycopg2==2.9.5
pyarrow==10.0.1
PyMySQL==1.0.2
pyodbc==4.0.34
pyparsing==3.0.9
PyPika==0.48.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
scipy==1.9.3
six==1.16.0
SQLAlchemy==1.4.43
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
